gemini_helper.py
import os
import requests
from dotenv import load_dotenv

load_dotenv()

API_KEY = os.getenv("GEMINI_API_KEY")
API_URL = f"https://generativelanguage.googleapis.com/v1beta/models/geminimodel?key={API_KEY}"
HEADERS = {"Content-Type": "application/json"}

def ask_gemini(prompt):
    body = {
        "contents": [
            {
                "parts": [{"text": prompt}]
            }
        ]
    }

    try:
        response = requests.post(API_URL, headers=HEADERS, json=body)
        response.raise_for_status()
        data = response.json()
        candidates = data.get("candidates", [])
        if candidates:
            return candidates[0]["content"]["parts"][0]["text"]
        else:
            return "⚠️ No response from Gemini."
    except requests.exceptions.RequestException as e:
        return f"❌ API Error: {e}"
